backend/requirements.txt
fastapi
uvicorn
yfinance
sqlalchemy
asyncpg
alembic
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv
requests
resend
lxml
